+++ examples/v0/example-goal.ts
