+++ examples/v0/logic.ts
