+++ packages/core/src/core/v1/memory.ts
@@ -1,4 +1,4 @@
-import { MemoryStore, WorkingMemory } from "../../../../../examples/new/types";
+import type { MemoryStore, WorkingMemory } from "./types";
 
 export async function getOrCreateConversationMemory(
     memory: MemoryStore,
@@ -18,16 +18,16 @@ export async function getOrCreateConversationMemory(
 export function createMemoryStore(): MemoryStore {
     const data = new Map<string, any>();
     return {
-        async get(key) {
+        async get(key: string) {
             return data.get(key);
         },
         async clear() {
             data.clear();
         },
-        async delete(key) {
+        async delete(key: string) {
             data.delete(key);
         },
-        async set(key, value) {
+        async set(key: string, value: any) {
             data.set(key, value);
         },
     };
